refactor: db update logic in OhlcvTable & moved some files
