--- src/data/database/__init__.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Dec 23 03:21:33 2024
-
-@author dhaneor
-"""
-from .base import DatabaseManager  # noqa: F401
-from .tables import OhlcvTable  # noqa: F401
